// src/components/Base/interface.ts
export enum ListLoadType {
  LIST_FIRST_JOIN = 'LIST_FIRST_JOIN',
  // export const LIST_NO_DATA = 'LIST_NO_DATA'
  LIST_LOAD_DATA = 'LIST_LOAD_DATA',
  LIST_LOAD_MORE = 'LIST_LOAD_MORE',
  LIST_LOAD_NO_MORE = 'LIST_LOAD_NO_MORE',
  LIST_LOAD_ERROR = 'LIST_LOAD_ERROR',
  LIST_NORMAL = 'LIST_NORMAL',
}